requirements.txt
Django>=3.2.12,<4.0
djangorestframework>=3.13.1,<3.14.0
drf-flex-fields>=0.9.8,<0.9.9
psycopg2>=2.8.6,<2.9
drf-spectacular>=0.22.1,<0.23
django-cors-headers>=3.12.0,<3.13.0